Assign array directly if type is unknown (#31)
With this change it is now possible to directly assign an array to its json/model when the type is unknown. This is useful when the structure of the array is dynamic or unknown.
In order to not break existing logic, a new configuration option `allow_generic_arrays` was added to the generator configuration.